Refactor utils (#541)
